requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorama==0.4.6
execnet==2.0.2
Faker==19.11.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
outcome==1.3.0
packaging==23.2
pip==22.3.1
pluggy==1.3.0
py==1.11.0
pycparser==2.21
PySocks==1.7.1
pytest-base-url==2.0.0
pytest-html==4.0.2
pytest-metadata==3.0.0
pytest-rerunfailures==12.0
pytest-selenium==4.0.1
pytest-variables==3.0.0
pytest-xdist==3.3.1
pytest==7.4.2
python-dateutil==2.8.2
requests==2.31.0
selenium==4.14.0
setuptools==65.5.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tenacity==8.2.3
trio-websocket==0.11.1
trio==0.22.2
urllib3==2.0.7
wheel==0.38.4
wsproto==1.2.0
